Messages: Message Should Not Be Constructed
`Message` is not a class that should be constructed. It's purpose is to
be a base class for `UserMessage` and `ClaudeMessage` and not to exist
on its own. This commit adds the `@abstractmethod` annotation to the
`Message.__init__` method so that it's not possible to instantiate a
`Message` on its own.