History: Add `as_strings` Method
This commit adds an `as_strings` method to the `History` class, along
with the appropriate unit tests. For use with the `create` API call, it is
necessary to be able to convert a `List[Message]`, which `History`
contains, into a `List[str]`, with the `str` here being the string
representation of a `Message`.