History: Add `as_dicts` Method to `History`
A prior commit mentioned the importance of structuring the `History` as
a `List[str]`. That was mistaken. It's actually important to structure
the historical messages as a `List[Dict[str, str]]` instead. This commit
adds an `as_dicts` method to `History` so that the returned result is in
the proper `List[Dict[str, str]]` format for use with `create` in the
API.